feat: MotionViewModel 프로퍼티 및 메서드 구현
- motionID추가
- 폴더 계층구조 변경
- MotionGraphViewModel -> MotionViewModel 이름 변경
- MotionPlayViewModel 삭제
