--- GyroData/Scene/MotionPlay/MotionPlayViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  MotionPlayViewModel.swift
-//  GyroData
-//
-//  Created by Ayaan, Wonbi on 2023/01/31.
-//
-
-import Foundation
